+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - **GM-982** - [Home] Fix background of the category icons
+- **GM-1002** - Created type `SearchState` for search
 
 ### Removed